Task #228554 -[FE] Beneficiary app folder Structure and naming,Task #228553-[FE] Benificiary App Layout
