+++ src/Screens/Profile.js
@@ -1,39 +1,38 @@
 import React from 'react';
 import {View, ScrollView, StyleSheet} from 'react-native';
-import Navbar from '../Component/Navbar';
-import CustomProgressBar from '../Component/CustomProgressBar';
-import DocumentList from '../Component/DocumentList';
+import Navbar from '../components/common/layout/Navbar';
+import DocumentList from '../components/DocumentList';
 import {documentName} from '../Constatnt/Constant';
-import HeadingText from '../Component/HeadingText';
-import CustomButton from '../Component/CustomButton';
+import HeadingText from '../components/common/layout/HeadingText';
+import CustomButton from '../components/common/button/Button';
 import {useNavigation} from '@react-navigation/native';
+import Layout from '../components/common/layout/Layout';
 const Profile = () => {
   const navigation = useNavigation();
   const handleOnPress = () => {
     navigation.navigate('Search');
   };
   return (
-    <View style={styles.container}>
-      <Navbar />
-      <HeadingText
-        benificiary={true}
-        heading={'Anaya Gupta'}
-        helperHeading={'Logged In With Digilocker'}
-      />
-      <CustomProgressBar totalDocuments={6} presentDocuments={6} />
+    <Layout
+      _heading={{
+        benificiary: true,
+        heading: 'Anaya Gupta',
+        subHeading: 'Logged in with digilocker',
+      }}>
+      {/* <CustomProgressBar totalDocuments={6} presentDocuments={6} /> */}
       {/* Scrollable part of the screen */}
       <ScrollView contentContainerStyle={styles.scrollContainer}>
         <DocumentList documents={documentName} />
         <CustomButton
           testId={'click-explore-scholarship'}
-          label="Explore Scholarship"
+          label="Explore Scholarships"
           marginTop={18}
           height={40}
           width={'95%'}
           handleClick={handleOnPress}
         />
       </ScrollView>
-    </View>
+    </Layout>
   );
 };
