+++ src/components/common/layout/SearchHeader.js
@@ -1,12 +1,12 @@
 import {StyleSheet, View} from 'react-native';
 import React from 'react';
-import FilterDialog from './FilterDialog';
-import CustomSearchbar from './CustomSearchbar';
+import FilterDialog from '../../FilterDialog';
+import Searchbar from '../inputs/Searchbar';
 
 const SearchHeader = () => {
   return (
     <View style={styles.view}>
-      <CustomSearchbar style={styles.searchBar} />
+      <Searchbar style={styles.searchBar} />
       <FilterDialog style={styles.filter} />
     </View>
   );
