+++ src/components/servises/benefits.js
@@ -1,0 +1,40 @@
+export const benefits = [
+  {
+    id: 1,
+    date: '12 September 2024',
+    title: 'Pre-Matric Scholarship-SC',
+    subTitle: 'Ministry of Social Justice',
+    benefitAmount: '3500 INR - 7700 INR',
+    eligible: ['SC', 'Madhya Pradesh', 'All Genders'],
+    description:
+      'The Pre-matric Scholarship-SC is provided by the Ministry of Social Justice and the Tribal Welfare Department of Madhya Pradesh. It supports Class 9 and 10 students from low-income families with an annual income below ₹2,50,000. Both genders are eligible.',
+    address: 'Madhya Pradesh',
+    imageUrl: 'https://via.placeholder.com/50',
+  },
+  {
+    id: 2,
+    date: '12 September 2024',
+    title: 'Pre-Matric Scholarship-ST',
+    subTitle: 'Ministry of Social Justice',
+    benefitAmount: '3500 INR - 7700 INR',
+
+    eligible: ['ST', 'Madhya Pradesh', 'All Genders'],
+    description:
+      'The Pre-matric Scholarship-ST is provided by the Ministry of Tribal Welfare and the Tribal Welfare Department of Madhya Pradesh. It supports Class 9 and 10 students from low-income families with an annual income below ₹2,50,000. Both male and female students are eligible.',
+    address: 'Madhya Pradesh',
+    imageUrl: 'https://via.placeholder.com/50',
+  },
+  {
+    id: 3,
+    date: '12 September 2024',
+    title: 'Pre-Matric Scholarship-ST',
+    subTitle: 'Ministry of Social Justice',
+    benefitAmount: '3500 INR - 7700 INR',
+
+    eligible: ['ST', 'Madhya Pradesh', 'All Genders'],
+    description:
+      'The Pre-matric Scholarship-ST is provided by the Ministry of Tribal Welfare and the Tribal Welfare Department of Madhya Pradesh. It supports Class 9 and 10 students from low-income families with an annual income below ₹2,50,000. Both male and female students are eligible.',
+    address: 'Madhya Pradesh',
+    imageUrl: 'https://via.placeholder.com/50',
+  },
+];
